qemuhotplugtest: add testcase to update a disk device
Since we can now use qemuDomainUpdateDeviceLiveAndConfig in tests,
updating more types of devices can be tested now.

One particular testcase was previously expected to trivially fail
because there was no way to test updating a disk device. Since now
it's possible to execute it, I modified it to make it work.
Message-Id: <1469735641-57657-6-git-send-email-t.flendrich@gmail.com>