device_conf: Let compiler decide on inlining functions
There's no point in forcing some functions to be both 'static'
and 'inline' at the same time at the header file level. This
leads to a situation where just linking the file defines those
function (potentially needlessly). Instead, we should let the
compiler (linker) decide whether the function are small enough to
be inlined or not.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <fb91867e2ee3ae58820b9eed3b7dc7356945a989.1470068509.git.mprivozn@redhat.com>