Introduce NVDIMM memory model
NVDIMM is new type of memory introduced in qemu. The idea is that
we have a DIMM module that keeps the data persistent across
domain reboots.
At the domain XML level, we already have some representation of
'dimm' modules. Long story short, we have <memory/> element that
lives under <devices/>. Now, the element even has @model
attribute which we can use to introduce new memory type:

    <memory model='nvdimm'>
      <source>
        <path>/tmp/nvdimm</path>
      </source>
      <target>
        <size unit='KiB'>523264</size>
        <node>0</node>
      </target>
    </memory>

So far, this is just a XML parser/formatter extension. QEMU
driver implementation is in the next commit.

For more info on NVDIMM visit the following web page:

    http://pmem.io/

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <e3ff7932116f8c7fd3a7e3fef698e0c071e329c8.1470921599.git.mprivozn@redhat.com>