qemu: Update guest CPU def in live XML
Storing the updated CPU definition in the live domain definition saves
us from having to update it over and over when we need it. Not to
mention that we will soon further update the CPU definition according to
QEMU once it's started.

A highly wanted side effect of this patch, libvirt will pass all CPU
features explicitly specified in domain XML to QEMU, even those that are
already included in the host model.

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>
Message-Id: <706cd9b89fb68c7abaaaba1a96976fae2009eb19.1471008154.git.jdenemar@redhat.com>