conf: Provide error on undefined vcpusched entry
Commit id '9cc931f0b' removed the error message; however, it was reachable
via the virsh edit of a domain and attempting to add a vcpusched element
for "undefined" vcpu values.  Without the error, the generic message
"An error occurred, but the cause is unknown" is displayed.

This is similar to bz 1366484 for the iothreadsched element.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1471272934-6529-3-git-send-email-jferlan@redhat.com>