qemu_hotplug: Support interface type of vhost-user hotplug
https://bugzilla.redhat.com/show_bug.cgi?id=1366108

So far, the hotplug of vhost-user "worked" by pure chance. Well,
qemu would error on our commands, but nevertheless. Now that we
have everything prepare, We should support hotplugging og
vhost-user. Firstly, we need to plug in chardev (through which
qemu talks to OpenVSwitch), then netdev and only after that we
can plug in the virtio-net-pci device.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <f4379a6a5c6dbeafa20c56988dc23c8b8c05d990.1471360664.git.mprivozn@redhat.com>