virlog: Introduce virLogFindOutput
Outputs are a bit trickier than filters, because if the user(config)-specified
set of outputs does contain duplicates. So, not only we would log twice, but
we would also leek FD for journald, since that one is global and is overwritten
every time a journald output was specified. For compatibility reasons, we
cannot just error out and forbid the daemon to start if we find duplicate
outputs which do not make sense. Instead, we could silently take into account
only the last occurrence of the duplicate output and remove all the previous
ones, so that the logger will not try to use them when it is looping over all
of its registered outputs.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <15cea16a343c5f957ec37bf777342ac2131374e3.1471519718.git.eskultet@redhat.com>