virlog: Introduce virLogParseOutput
Introduce a method to parse an individual logging output. The difference
compared to the virLogParseAndDefineOutput is that this method does not define
the output, instead it makes use of the virLogNewOutputTo* methods introduced
in the previous patch and just returns the virLogOutput object that has to be
added to a list of object which then can be defined as a whole via
virLogDefineOutputs. The idea remains still the same - split parsing and
defining of the logging primitives (outputs, filters).

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <7a1628a9e0b35495fb357b990c74932bff1be861.1471519718.git.eskultet@redhat.com>