virlog: Make some of the methods static
Several methods that were introduced in previous patches were not made static
on purpose so they could be introduced without any occurrence without the
compiler complaining about not using them. Since the scope of the methods is
local to the virlog module only, they can now be turned to static.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <91be2a9f2be8db7122597e04e5e4b7ff3d99ee13.1471519718.git.eskultet@redhat.com>