virlog: Store the journald fd within the output object
Now that previous patches tried to refactor the code and split parsing and
defining logic of logging primitives, there is no reason why we could not keep
journald's fd within the journald output object the same way as we do for
regular file-based outputs. Quite the opposite, by doing that we gain the
benefit of creating a journald-based object the same way as we create
file-based ones and therefore when replacing the existing set of outputs
for a new one, journald does not need to be special cased due to its globally
shared fd and the only remaining thing to special case is syslog. Additionally,
making this change, we don't need the virLogCloseJournald routinei anymore,
since as it stores the fd within the object, plain virLogCloseFd will suffice.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <4eaef04a8ccc9935f7346340bb1067f5e35d226a.1471519718.git.eskultet@redhat.com>