qemu: Use mondern vcpu hotplug approach if possible
To allow unplugging the vcpus, hotplugging of vcpus on platforms which
require to plug multiple logical vcpus at once or pluging them in in
arbitrary order it's necessary to use the new device_add interface for
vcpu hotplug.

This patch adds support for the device_add interface using the old
setvcpus API by implementing an algorihm to select the appropriate
entities to plug in.
Message-Id: <d3e002f2af43b8b0ac551788b4e029fda0572542.1471616971.git.pkrempa@redhat.com>