Introduce a "scsi_host" hostdev type
We already have a "scsi" hostdev type, which refers to a single LUN
that is passed through to a guest.  But what of things where multiple
LUNs are passed through via a single SCSI HBA, such as with the
vhost-scsi target?  Create a new hostdev type that will carry
this, and its associated documentation and XML schema information.

Signed-off-by: Eric Farman <farman@linux.vnet.ibm.com>
Message-Id: <1473166712-29674-2-git-send-email-farman@linux.vnet.ibm.com>